[Swift in WebKit] Non-PAL targets should not access the internal PAL Swift bridging header (part 4) https://bugs.webkit.org/show_bug.cgi?id=310681 rdar://173281825
Reviewed by Aditya Keerthi.

Continuing working on fixing the layering violations by moving more stuff into PAL.

* Source/WebCore/PAL/PAL.xcodeproj/project.pbxproj:
* Source/WebCore/PAL/pal/crypto/CryptoAlgorithmAESKWCocoaBridging.cpp: Copied from Source/WebCore/crypto/cocoa/CryptoAlgorithmAESKWCocoa.cpp.
(PAL::Crypto::wrapKeyAESKWCryptoKit):
(PAL::Crypto::unwrapKeyAESKWCryptoKit):
* Source/WebCore/PAL/pal/crypto/CryptoAlgorithmAESKWCocoaBridging.h: Copied from Source/WebCore/PAL/pal/crypto/CryptoTypes.h.
* Source/WebCore/PAL/pal/crypto/CryptoAlgorithmEd25519CocoaBridging.cpp: Copied from Source/WebCore/crypto/cocoa/CryptoAlgorithmEd25519Cocoa.cpp.
(PAL::Crypto::signEd25519CryptoKit):
(PAL::Crypto::verifyEd25519CryptoKit):
* Source/WebCore/PAL/pal/crypto/CryptoAlgorithmEd25519CocoaBridging.h: Copied from Source/WebCore/PAL/pal/crypto/CryptoTypes.h.
* Source/WebCore/PAL/pal/crypto/CryptoAlgorithmX25519CocoaBridging.cpp: Copied from Source/WebCore/crypto/cocoa/CryptoDigestAlgorithm.h.
(PAL::Crypto::deriveBitsCryptoKit):
* Source/WebCore/PAL/pal/crypto/CryptoAlgorithmX25519CocoaBridging.h: Copied from Source/WebCore/PAL/pal/crypto/CryptoTypes.h.
* Source/WebCore/PAL/pal/crypto/CryptoTypes.h:

- Sink these functions into PAL

* Source/WebCore/PAL/pal/module.modulemap:

- Workaround for module incremental build issue

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:

- Add the files to the Xcode project

* Source/WebCore/crypto/CryptoAlgorithmIdentifier.h:
(WebCore::toCKHashFunction): Deleted.
* Source/WebCore/crypto/cocoa/CryptoTypesBridging.h: Renamed from Source/WebCore/crypto/cocoa/CryptoDigestAlgorithm.h.
(WebCore::cryptoDigestAlgorithm):
(WebCore::toCKHashFunction):
(WebCore::toExceptionCode):
(WebCore::toException):
* Source/WebCore/crypto/algorithms/CryptoAlgorithmEd25519.h:
* Source/WebCore/crypto/cocoa/CryptoAlgorithmAESGCMCocoa.cpp:
(WebCore::toExceptionCode): Deleted.
(WebCore::toException): Deleted.

- Combine into a single header

* Source/WebCore/crypto/cocoa/CryptoAlgorithmAESKWCocoa.cpp:
(WebCore::CryptoAlgorithmAESKW::platformWrapKey):
(WebCore::CryptoAlgorithmAESKW::platformUnwrapKey):
(WebCore::wrapKeyAESKWCryptoKit): Deleted.
(WebCore::unwrapKeyAESKWCryptoKit): Deleted.
* Source/WebCore/crypto/cocoa/CryptoAlgorithmECDSACocoa.cpp:
* Source/WebCore/crypto/cocoa/CryptoAlgorithmEd25519Cocoa.cpp:
(WebCore::CryptoAlgorithmEd25519::platformSign):
(WebCore::CryptoAlgorithmEd25519::platformVerify):
(WebCore::signEd25519CryptoKit): Deleted.
(WebCore::verifyEd25519CryptoKit): Deleted.
* Source/WebCore/crypto/cocoa/CryptoAlgorithmHKDFCocoa.cpp:
* Source/WebCore/crypto/cocoa/CryptoAlgorithmHMACCocoa.cpp:
* Source/WebCore/crypto/cocoa/CryptoAlgorithmRSASSA_PKCS1_v1_5Cocoa.cpp:
* Source/WebCore/crypto/cocoa/CryptoAlgorithmRSA_PSSCocoa.cpp:
* Source/WebCore/crypto/cocoa/CryptoAlgorithmX25519Cocoa.cpp:
(WebCore::CryptoAlgorithmX25519::platformDeriveBits):
(WebCore::deriveBitsCryptoKit): Deleted.
* Source/WebCore/crypto/cocoa/CryptoKeyECCocoa.cpp:

- Update imports and call sites

Canonical link: https://commits.webkit.org/309892@main
