+++ Source/WebCore/PAL/pal/module.modulemap
@@ -29,4 +29,4 @@ module pal [system] {
     export *
 }
 
-// (v1) This comment ensures an incremental build causes this module to rebuild (rdar://170129992)
+// (v2) This comment ensures an incremental build causes this module to rebuild (rdar://170129992)
