+++ Source/WebCore/crypto/cocoa/CryptoAlgorithmAESKWCocoa.cpp
@@ -28,39 +28,19 @@
 
 #include "CryptoKeyAES.h"
 #include <CommonCrypto/CommonCrypto.h>
+#include <pal/crypto/CryptoAlgorithmAESKWCocoaBridging.h>
 #include <pal/crypto/CryptoTypes.h>
-
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wunsafe-buffer-usage"
-#include "PALSwift-Generated.h"
-#pragma clang diagnostic pop
 
 namespace WebCore {
 
-static ExceptionOr<Vector<uint8_t>> wrapKeyAESKWCryptoKit(const Vector<uint8_t>& key, const Vector<uint8_t>& data)
-{
-    auto rv = pal::AesKw::wrap(data.span(), key.span());
-    if (rv.errorCode != PAL::Crypto::Error::Success)
-        return Exception { ExceptionCode::OperationError };
-    return WTF::move(rv.result);
-}
-
-static ExceptionOr<Vector<uint8_t>> unwrapKeyAESKWCryptoKit(const Vector<uint8_t>& key, const Vector<uint8_t>& data)
-{
-    auto rv = pal::AesKw::unwrap(data.span(), key.span());
-    if (rv.errorCode != PAL::Crypto::Error::Success)
-        return Exception { ExceptionCode::OperationError };
-    return WTF::move(rv.result);
-}
-
 ExceptionOr<Vector<uint8_t>> CryptoAlgorithmAESKW::platformWrapKey(const CryptoKeyAES& key, const Vector<uint8_t>& data)
 {
-    return wrapKeyAESKWCryptoKit(key.key(), data);
+    return toException(PAL::Crypto::wrapKeyAESKWCryptoKit(key.key(), data));
 }
 
 ExceptionOr<Vector<uint8_t>> CryptoAlgorithmAESKW::platformUnwrapKey(const CryptoKeyAES& key, const Vector<uint8_t>& data)
 {
-    return unwrapKeyAESKWCryptoKit(key.key(), data);
+    return toException(PAL::Crypto::unwrapKeyAESKWCryptoKit(key.key(), data));
 }
 
 } // namespace WebCore
