+++ Source/WebCore/crypto/cocoa/CryptoAlgorithmEd25519Cocoa.cpp
@@ -28,42 +28,20 @@
 
 #include "CryptoKeyOKP.h"
 #include "ExceptionOr.h"
+#include <pal/crypto/CryptoAlgorithmEd25519CocoaBridging.h>
 #include <pal/crypto/CryptoTypes.h>
 #include <pal/spi/cocoa/CoreCryptoSPI.h>
 
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wunsafe-buffer-usage"
-#include "PALSwift-Generated.h"
-#pragma clang diagnostic pop
-
 namespace WebCore {
-
-static ExceptionOr<Vector<uint8_t>> signEd25519CryptoKit(const Vector<uint8_t>&sk, const Vector<uint8_t>& data)
-{
-    if (sk.size() != ed25519KeySize)
-        return Exception { ExceptionCode::OperationError };
-    auto rv = pal::EdKey::sign(pal::EdSigningAlgorithm::ed25519(), sk.span(), data.span());
-    if (rv.errorCode != PAL::Crypto::Error::Success)
-        return Exception { ExceptionCode::OperationError };
-    return WTF::move(rv.result);
-}
-
-static ExceptionOr<bool> verifyEd25519CryptoKit(const Vector<uint8_t>& pubKey, const Vector<uint8_t>& signature, const Vector<uint8_t>& data)
-{
-    if (pubKey.size() != ed25519KeySize || signature.size() != ed25519SignatureSize)
-        return false;
-    auto rv = pal::EdKey::verify(pal::EdSigningAlgorithm::ed25519(), pubKey.span(), signature.span(), data.span());
-    return rv.errorCode == PAL::Crypto::Error::Success;
-}
 
 ExceptionOr<Vector<uint8_t>> CryptoAlgorithmEd25519::platformSign(const CryptoKeyOKP& key, const Vector<uint8_t>& data)
 {
-    return signEd25519CryptoKit(key.platformKey(), data);
+    return toException(PAL::Crypto::signEd25519CryptoKit(key.platformKey(), data));
 }
 
 ExceptionOr<bool> CryptoAlgorithmEd25519::platformVerify(const CryptoKeyOKP& key, const Vector<uint8_t>& signature, const Vector<uint8_t>& data)
 {
-    return verifyEd25519CryptoKit(key.platformKey(), signature, data);
+    return toException(PAL::Crypto::verifyEd25519CryptoKit(key.platformKey(), signature, data));
 }
 
 } // namespace WebCore
