+++ Source/WebCore/crypto/cocoa/CryptoTypesBridging.h
@@ -1,0 +1,85 @@
+/*
+ * Copyright (C) 2017 Apple Inc. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS''
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS
+ * BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
+ * THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#pragma once
+
+#include <WebCore/CryptoAlgorithmIdentifier.h>
+#include <WebCore/ExceptionOr.h>
+#include <optional>
+#include <pal/crypto/CryptoTypes.h>
+#include <wtf/Vector.h>
+
+namespace WebCore {
+
+inline std::optional<PAL::Crypto::CryptoDigest::Algorithm> cryptoDigestAlgorithm(CryptoAlgorithmIdentifier hashFunction)
+{
+    switch (hashFunction) {
+    case CryptoAlgorithmIdentifier::SHA_1:
+        return PAL::Crypto::CryptoDigest::Algorithm::SHA_1;
+    case CryptoAlgorithmIdentifier::SHA_256:
+        return PAL::Crypto::CryptoDigest::Algorithm::SHA_256;
+    case CryptoAlgorithmIdentifier::SHA_384:
+        return PAL::Crypto::CryptoDigest::Algorithm::SHA_384;
+    case CryptoAlgorithmIdentifier::SHA_512:
+        return PAL::Crypto::CryptoDigest::Algorithm::SHA_512;
+    default:
+        return std::nullopt;
+    }
+}
+
+inline PAL::Crypto::CryptoDigestHashFunction toCKHashFunction(CryptoAlgorithmIdentifier hash)
+{
+    if (auto result = cryptoDigestAlgorithm(hash))
+        return *result;
+
+    ASSERT_NOT_REACHED();
+    return PAL::Crypto::CryptoDigestHashFunction::SHA_512;
+}
+
+inline std::optional<ExceptionCode> toExceptionCode(const PAL::Crypto::Error& error)
+{
+    switch (error) {
+    case PAL::Crypto::Error::Success:
+        return std::nullopt;
+
+    default:
+        return ExceptionCode::OperationError;
+    }
+}
+
+template<typename T>
+inline ExceptionOr<T> toException(Expected<T, PAL::Crypto::Error>&& expected)
+{
+    if (expected)
+        return WTF::move(*expected);
+
+    auto exceptionCode = toExceptionCode(expected.error());
+    if (exceptionCode)
+        return Exception { *exceptionCode };
+
+    RELEASE_ASSERT_NOT_REACHED();
+}
+
+} // namespace WebCore
